[client,android] Remove QuickConnectBookmark
