+++ client/Android/Studio/freeRDPCore/src/main/java/com/freerdp/freerdpcore/services/QuickConnectHistoryGateway.java
@@ -13,7 +13,7 @@ package com.freerdp.freerdpcore.services;
 import com.freerdp.freerdpcore.data.HistoryDao;
 import com.freerdp.freerdpcore.data.HistoryEntity;
 import com.freerdp.freerdpcore.domain.BookmarkBase;
-import com.freerdp.freerdpcore.domain.QuickConnectBookmark;
+import com.freerdp.freerdpcore.domain.ManualBookmark;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -35,7 +35,7 @@ public class QuickConnectHistoryGateway
 		ArrayList<BookmarkBase> result = new ArrayList<>(entities.size());
 		for (HistoryEntity entity : entities)
 		{
-			QuickConnectBookmark bookmark = new QuickConnectBookmark();
+			ManualBookmark bookmark = new ManualBookmark();
 			bookmark.setLabel(entity.item);
 			bookmark.setHostname(entity.item);
 			result.add(bookmark);
